fixed untracked files
